Add ODC-BY Open Data LICENSE
Open Data Commons Attribution License (ODC-By) v1.0

This is a human-readable summary of the ODC-BY 1.0 license. Please see the disclaimer below.

You are free:

To share: To copy, distribute and use the database.
To create: To produce works from the database.
To adapt: To modify, transform and build upon the database.
As long as you:

Attribute: You must attribute any public use of the database, or works produced from the database, in the manner specified in the license. For any use or redistribution of the database, or works produced from it, you must make clear to others the license of the database and keep intact any notices on the original database.